requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
filelock==3.7.1
huggingface-hub==0.7.0
idna==3.3
jiwer==2.3.0
numpy==1.22.4
packaging==21.3
pydub==0.25.1
pyparsing==3.0.9
python-Levenshtein==0.12.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
sqlparse==0.4.2
tokenizers==0.12.1
torch==1.11.0
torchaudio==0.11.0
touch==2020.12.3
tqdm==4.64.0
transformers==4.19.4
typing_extensions==4.2.0
urllib3==1.26.9
values==2020.12.3
